core/types: Move receipt tests OP diff to separate file
